requirements.txt
fastapi
uvicorn[standard]
python-dotenv
pydantic
pydantic-settings
SQLAlchemy
python-jose[cryptography]
passlib[bcrypt]
python-multipart